+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 - Add head_js block into base html template
 - list teacher's organizations in the teacher dashboard sidebar.
 - Added Certificates in the dashboard
+- Dashboard infinite scroll of orders and enrollments
 
 ### Fixed